remove old scarunner code and move its tests to the right place
